# docs/swagger.yaml
basePath: /
definitions:
...
  api.HttpError:
    properties:
      error:
        type: string
      statusCode:
        type: integer
    type: object
...
  api.StatsReset:
    properties:
      games:
        type: integer
      guildId:
        type: string
      userId:
        type: string
    type: object
...
  game.GameDelays:
    properties:
      delays:
        additionalProperties:
          additionalProperties:
            type: integer
          type: object
        description: maps from origin->new phases, with the integer number of seconds
          for the delay
        type: object
    type: object
  game.VoiceRules:
    properties:
      deafRules:
        additionalProperties:
          additionalProperties:
            type: boolean
          type: object
        type: object
      muteRules:
        additionalProperties:
          additionalProperties:
            type: boolean
          type: object
        type: object
    type: object
...
  settings.GuildSettings:
    properties:
      adminIDs:
        items:
          type: string
        type: array
      autoRefresh:
        type: boolean
      delays:
        $ref: '#/definitions/game.GameDelays'
      deleteGameSummary:
        type: integer
      displayRoomCode:
        type: string
      language:
        type: string
      leaderboardMention:
        type: boolean
      leaderboardMin:
        type: integer
      leaderboardSize:
        type: integer
      mapVersion:
        type: string
      matchSummaryChannelID:
        type: string
      muteSpectator:
        type: boolean
      permissionRoleIDs:
        items:
          type: string
        type: array
      unmuteDeadDuringTasks:
        type: boolean
      voiceRules:
        $ref: '#/definitions/game.VoiceRules'
    type: object
info:
  contact: {}
  description: AutoMuteUs API, served independently of Discord bot shards.
  title: AutoMuteUs
  version: dev
paths:
...
  /guild/settings/reset:
    post:
      description: |-
        Put every setting of the guild back to its default, like the /settings reset slash command. This
        includes the match summary channel and operator roles. Premium is not checked, since the defaults
        need none. Requires the guild owner or the Discord Administrator or Manage Server permission, and
        counts against the same per-guild budget as PATCH /guild/settings.
      parameters:
      - description: Guild ID
        in: query
        name: guildID
        required: true
        type: string
      - description: ETag from a previous GET or PATCH; the reset is refused with
          412 if the settings changed since
        in: header
        name: If-Match
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The default settings, now stored
          headers:
            ETag:
              description: Version of the stored settings after the reset
              type: string
          schema:
            $ref: '#/definitions/settings.GuildSettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HttpError'
        "409":
          description: Settings were changed concurrently; reload and retry
          schema:
            $ref: '#/definitions/api.HttpError'
        "412":
          description: If-Match did not match the stored version
          schema:
            $ref: '#/definitions/api.HttpError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/api.HttpError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.HttpError'
      security:
      - BasicAuth: []
      - DiscordBearer: []
      summary: Reset Guild Settings
      tags:
      - guild
...
  /guild/stats/reset:
    post:
      description: |-
        Delete every recorded game of the guild, for every player, like the /stats guild reset slash command.
        It cannot be undone. Requires the guild owner or the Discord Administrator or Manage Server
        permission.
      parameters:
      - description: Guild ID
        in: query
        name: guildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatsReset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpError'
      security:
      - BasicAuth: []
      - DiscordBearer: []
      summary: Reset Guild Stats
      tags:
      - guild
...
  /guild/user/reset:
    post:
      description: |-
        Remove one player from every recorded game of this guild. Their games in other guilds are kept, and
        so are the games themselves, so the other players' stats do not change. It cannot be undone.
        Requires the guild owner or the Discord Administrator or Manage Server permission, including to
        reset one's own stats.
      parameters:
      - description: Guild ID
        in: query
        name: guildID
        required: true
        type: string
      - description: User ID
        in: query
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatsReset'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpError'
      security:
      - BasicAuth: []
      - DiscordBearer: []
      summary: Reset Player Stats
      tags:
      - guild
...
securityDefinitions:
  BasicAuth:
    type: basic
  DiscordBearer:
    description: 'Discord OAuth access token: Bearer <token>; requires identify and
      guilds.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"